Split up uifunc library to make testing easier
Move widget-related functions to widgetfunc
Replace some calls to uifunc::stopApp() with errorfunc::fatalError()
Move alert when QMediaPlayer fails to create from soundfunc to calling code
