+++ docs/source/autodoc/tablet_qt/_index.rst
@@ -224,6 +224,8 @@ tablet_qt
     lib/datetime.h.rst
     lib/debugfunc.cpp.rst
     lib/debugfunc.h.rst
+    lib/errorfunc.cpp.rst
+    lib/errorfunc.h.rst
     lib/filefunc.cpp.rst
     lib/filefunc.h.rst
     lib/flagguard.cpp.rst
@@ -258,6 +260,8 @@ tablet_qt
     lib/uifunc.h.rst
     lib/version.cpp.rst
     lib/version.h.rst
+    lib/widgetfunc.cpp.rst
+    lib/widgetfunc.h.rst
     main.cpp.rst
     maths/ccrandom.cpp.rst
     maths/ccrandom.h.rst
