+++ tablet_qt/dialogs/soundtestdialog.cpp
@@ -19,6 +19,7 @@
 */
 
 #include "soundtestdialog.h"
+#include "common/textconst.h"
 #include "dialogs/logbox.h"
 #include "lib/soundfunc.h"
 #include "lib/uifunc.h"
@@ -30,6 +31,7 @@ SoundTestDialog::SoundTestDialog(const QUrl& url, const int volume_percent,
 {
     soundfunc::makeMediaPlayer(m_player);
     if (!m_player) {
+        uifunc::alert(TextConst::unableToCreateMediaPlayer());
         return;
     }
     connect(m_player.data(), &QMediaPlayer::mediaStatusChanged,
