+++ tablet_qt/lib/soundfunc.cpp
@@ -23,7 +23,6 @@
 #include <QObject>
 #include <QMediaPlayer>
 #include "common/textconst.h"
-#include "lib/uifunc.h"
 #include "maths/mathfunc.h"
 
 namespace soundfunc {
@@ -101,10 +100,6 @@ void makeMediaPlayer(QSharedPointer<QMediaPlayer>& player)
         https://launchpad.net/distros/ubuntu/+source/gstreamer1.0
 
     */
-
-    if (!player) {
-        uifunc::alert(TextConst::unableToCreateMediaPlayer());
-    }
 }
