+++ tablet_qt/lib/uifunc.h
@@ -109,57 +109,6 @@ QAbstractButton* iconButton(const QString& normal_filename,
                             const QString& pressed_filename = QString(),
                             QWidget* parent = nullptr);
 
-// ============================================================================
-// Widget manipulations, and other Qt internals
-// ============================================================================
-
-// Set a widget's background colour by setting its CSS stylesheet.
-void setBackgroundColour(QWidget* widget, const QColor& colour);
-
-// Set a widget's background colour and "pressed" background colour
-// by setting its CSS stylesheet.
-void setBackgroundAndPressedColour(QWidget* widget,
-                                   const QColor& background,
-                                   const QColor& pressed);
-
-// Delete all children of a widget.
-void removeAllChildWidgets(QObject* object);
-
-// Combines a horizontal and a vertical alignment into an alignment object
-// that carries the same information jointly.
-Qt::Alignment combineAlignment(Qt::Alignment halign, Qt::Alignment valign);
-
-// Repolishes the widget.  Calls the widget style's unpolish() then polish()
-// functions, then the widget's update().
-void repolish(QWidget* widget);
-
-// Calls the widget's setProperty() with a property name/value pair, converting
-// to the requisite Qt types on the way. Optionally, repolishes the widget.
-void setProperty(QWidget* widget, const QString& property,
-                 const QVariant& value, bool repolish = true);
-
-// Converts a bool to "true" or "false" (for use in CSS).
-QString cssBoolean(bool value);
-
-// And some specific ones:
-
-// Sets the widget's "italic" property.
-void setPropertyItalic(QWidget* widget, bool italic, bool repolish = true);
-
-// Sets the widget's "missing" property.
-void setPropertyMissing(QWidget* widget, bool missing,
-                        bool repolish = true);
-
-// Clear all widgets from a layout
-void clearLayout(QLayout* layout, bool delete_widgets = true);
-
-// Handy functions:
-
-// Scrolls an editor to the end (bottom left).
-void scrollToEnd(QPlainTextEdit* editor);
-
-// Scrolls an editor to the top (top left).
-void scrollToStart(QPlainTextEdit* editor);
 
 // ============================================================================
 // Killing the app
@@ -258,10 +207,6 @@ void visitUrl(const QString& url);
 // ============================================================================
 // Strings
 // ============================================================================
-
-// Escapes a string to a double-quoted C++ literal.
-// (Some DUPLICATION; see convert.h!)
-QString escapeString(const QString& string);
 
 // Converts a bool to a tr()-translated "Yes" or "No".
 QString yesNo(bool yes);
