+++ tablet_qt/questionnairelib/qucountdown.cpp
@@ -28,6 +28,7 @@
 #include <QPushButton>
 #include <QTimer>
 #include "common/cssconst.h"
+#include "common/textconst.h"
 #include "lib/soundfunc.h"
 #include "lib/uifunc.h"
 #include "questionnairelib/questionnaire.h"
@@ -108,6 +109,8 @@ QPointer<QWidget> QuCountdown::makeWidget(Questionnaire* questionnaire)
         soundfunc::makeMediaPlayer(m_player);
         if (m_player) {
             m_player->setMedia(QUrl(uiconst::SOUND_COUNTDOWN_FINISHED));
+        } else {
+            uifunc::alert(TextConst::unableToCreateMediaPlayer());
         }
     }
