+++ tablet_qt/questionnairelib/quspinboxdouble.cpp
@@ -22,7 +22,7 @@
 
 #include "quspinboxdouble.h"
 #include <QDoubleSpinBox>
-#include "lib/uifunc.h"
+#include "lib/widgetfunc.h"
 #include "questionnaire.h"
 
 
@@ -113,7 +113,7 @@ void QuSpinBoxDouble::fieldValueChanged(const FieldRef* fieldref,
     if (!m_spinbox) {
         return;
     }
-    uifunc::setPropertyMissing(m_spinbox, fieldref->missingInput());
+    widgetfunc::setPropertyMissing(m_spinbox, fieldref->missingInput());
     if (originator != this) {
         const QSignalBlocker blocker(m_spinbox);
         m_spinbox->setValue(fieldref->valueDouble());
